backend/requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic[email]
python-multipart

# Data processing
beautifulsoup4>=4.9.3
newspaper3k>=0.2.8
python-dateutil>=2.8.2
PyMuPDF>=1.18.0
youtube_transcript_api>=0.4.1
arxiv>=1.4.2
scholarly>=1.7.0
crossref-commons>=0.0.7
tiktoken>=0.3.0

# Machine Learning
transformers
sentence-transformers
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2

# Vector Database & LLM
pathway[xpack-llm]==0.16.0
llama-index-core>=0.10.0
llama-index-retrievers-pathway>=0.1.0

# LLM Dependencies
litellm>=1.30.0  # For Groq and OpenRouter support
anthropic>=0.18.1  # For Claude models
pathway[xpack-llm]>=0.8.0
python-dotenv>=1.0.0

# Utilities
tenacity>=8.0.1
cryptography
python-jose
passlib
lxml[html_clean]

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5

langchain
anthropic==0.18.1
python-dotenv==1.0.0
mistralai==0.0.12
typing>=3.7.4
groq==0.4.2
llama-index-core>=0.10.0
pydantic
aiohttp>=3.9.0
PyMuPDF>=1.23.0
arxiv>=2.1.0
tenacity>=8.2.0
youtube-transcript-api>=0.6.2
requests>=2.31.0
newspaper3k>=0.2.8
langchain-core
langchain-anthropic
asyncio>=3.4.3
duckduckgo-search==6.3.7
primp==0.8.1